QuickSpec: report failures via exceptions
Prior to this commit, Nimble would report failures via XCTFail. While
this allows for a complete separation of Quick and Nimble, it is
inflexible--if expectations are placed within a shared example group,
for example, Quick has no way of reporting the failures accurately.

Quick now reports failures via exceptions. Nimble raises exceptions to
indicate failures, with details on the file and line number of the
failed expectation embedded within the raised exception's user info.

This paves the way for shared examples #44 and
custom output formatters #9.

Other changes of note:

- Remove API to override failure reporting in QuickSpec. Since QuickSpec
  uses SenTestingKit file and line number keys anyway, the issue raised
  in #41 can still be considered resolved
- Rename Callsite_ to Callsite--I don't know how I missed this in the
  pull request review