Quick.DSL: add shared examples
Allow users to define shared examples using `sharedExamples`, and to
inject those examples anywhere using `itBehavesLike`.

Shared examples are registered with the World. Attempts to register
shared examples with the same name throw an exception. An attempt to
execute shared examples that have not been registered also raises an
exception.

`itBehaveLike` acts like a `describe` block, except that if
any of the examples in an `itBehaveLike` fails, the line where
`itBehaveLike` is executed is reported as an XCTest failure, not the
line with the failing expectation.