Quick.DSL: shared examples may use extra context
`itBehavesLike` now takes an extra, optional argument: a closure that,
when evaluated, returns a hash of values that may be used in shared
examples.

`sharedExamples` may be passed either a closure that takes no arguments,
or one that uses these additional hash values.

Completes issue #44, which in turn resolves issue #37.