Remove Quick.xcworkspace
Quick.xcworkspace contained both the Quick and Nimble projects. This
used to be useful, since both projects were changing drastically from
day to day, and it was nice to be able to edit both easily.

Now that things have calmed down, it's silly for the Quick repo to
contain infrastucture to build and test Nimble.

- Remove Quick.xcworkspace
- Move Quick.xcodeproj to the root directory
- `rake test` now only runs Quick tests