Example: Don't include internal root group in name
Due to an internal implementation detail, all examples belong to an
example group. When not explicitly added to an example group, Quick adds
examples to an internal "root example group". This is to handle cases
like:

    override func spec() {
      beforeEach { /* ... */ }
      it("is an example") { /* ... */ }
    }

`beforeEach` closures need to be executed within an example group, so in
the above case Quick adds the example to an internal group.

This internal group, however, appears in example names. The above
example name was reported as "root example group, is an example". This
obviously isn't the intended example name, and the bug was reported in
issue #168.

In order to fix the issue, add a flag to example groups to determine
whether they are internal root example groups. If they are, do not
return a name.