Move Nimble dependency to a workspace
This creates a workspace to hold both the Quick and Nimble projects.
Nimble is a dependency for the tests, but isn't needed in the main
framework targets.

Having Nimble embedded in Quick.xcodeproj was causing issues with
[Carthage] building both Quick and Nimble. The end result of the issue
was that if you included both Quick and Nimble, only the version of
Nimble included with Quick would exist after compilation. By moving it
into a workspace, we can ensure that Carthage will only build the Quick
framework, letting the user manage Nimble separately.

[Carthage]: https://github.com/Carthage/Carthage