AfterEachTests: Test order
Functional tests for afterEach tested the total number of times
afterEach closures were called, but not the order in which they were
called. Test the order as well.

This is brittle in that these tests may break when test order shuffling
is introduced (see #152)--but when that time comes, shuffling should
simply be disabled when running these order-based tests.