README: Move documentation into separate directory
The README currently contains information on what Quick is, why you'd
want to use Quick, how to install Quick, and more. This makes it
extremely long, and difficult to read.

Split the README up into a bunch of separate Markdown files, compiled in
a directory called `Documentation`. The documentation directory mostly
contains instructions on using Quick. However, since some aspects of
Quick are best appreciated when one has some knowledge of how to write
unit tests in Swift/Objective-C, the documentation has instructions on
that as well.