QCKDSL.h: Bridge beforeEach with metadata to Obj-C
A goal for Quick v1.0 is to no longer expose the Quick.World singleton.
This is made harder by the fact that our very own tests use it!

Bridge beforeEach closures that take example metadata arguments to
Objective-C, in order to allow ItTests+ObjC.m to not use "private"
APIs.