QuickTests: Suffix Objective-C tests with "_ObjC"
When investigating #373, I ran into an issue in which
`+[XCTestCase testInvocations]` would return the **wrong set of
invocations** when two test cases shared the same class name!

This sounds absolutely insane, but it's true! Notice that if you define
two test cases, one in Swift and one in Objective-C, but with the same
name, their test methods are grouped in Xcode's test navigator:

http://f.cl.ly/items/17412d0p291d0p0H2u30/Screen%20Shot%202015-09-23%20at%2011.03.03%20PM.png

The test case class names don't collide when defined in Swift and
Objective-C respectively, but end up behaving very erratically. To
prevent this issue, make sure Objective-C and Swift tests are named
"namespaced" by affixing "_ObjC" to Objective-C test classes.

This probably deserves an Apple radar.