Remove public headers from umbrella header
This seems to be the root cause for new projects created in Xcode 7.1
to report "Include of non-modular header inside framework Module
'Quick'".

Infer: It seems like the swift compiler in Xcode 7.1 now prevents
non-modular includes in umbrella header files. The public headers are
assumed to be imported in modules.