[Tests] Remove old ContextTests file
The file was copied to a new location as part of https://github.com/Quick/Quick/pull/448,
but the file at the old location wasn't removed. Remove the old file,
which is no longer used in any way.

As a result, the entire `QuickTests` directory no longer exists in the
repository.