Cache the included example count on Example.run()
Due to the fact that world.includedExamples is a computed property that
does  a filter on all the tests, performing
`numberOfExamplesRun >= world.includedExampleCount` is not very efficient
with a very large number of tests.

This patch caches the result of world.includedExampleCount and uses the
cached value to perform the comparison.

Fix #696