Add support for running a single test
This implements half of the functionality of [zacwest's PR](https://github.com/Quick/Quick/pull/936), but in a more maintainable manner.
This allows you to select an individual (or set of) test from the Test Navigator and run only that test.

This also adds tests guaranteeing the functionality works, though it is split between tests written in objc and swift.