Add async resultbuilder-based support
This basically serves to reimplement Quick using a ResultBuilder approach.
It doesn't integrate with anything else in Quick, and I only provide basic beforeEach, it (the actual test), afterEach, as well as example groups.

This should serve as a proof of concept for this approach.
I want to test out how this scales before fully committing to it.

Assuming we like this, then I'll commit to adding other features of the DSL to this.